+++ src/game/larafire.c
@@ -628,7 +628,7 @@ void HitTarget(ITEM_INFO *item, GAME_VECTOR *hitpos, int32_t damage)
             Sound_Effect(SFX_RAT_CHIRP, &item->pos, SPM_NORMAL);
             break;
 
-        case O_MERCENARY1:
+        case O_SKATEKID:
             Sound_Effect(SFX_SKATEBOARD_HIT, &item->pos, SPM_NORMAL);
             break;